w3c-reports/12092014/gtk/runned-tests/touch-events/create-touch-touchlist-expected.txt

Touch Events: createTouch and createTouchList tests

Test complete.

PASS document.createTouch exists and creates a Touch object with requested properties 
PASS document.createTouchList exists and correctly creates a TouchList from zero Touch objects 
PASS touch list is a TouchList object 
PASS TouchList.length attribute exists 
PASS TouchList.length attribute is of type number (unsigned long) 
PASS TouchList.item attribute exists 
PASS TouchList.item attribute is of type function 
PASS TouchList.item returns null if the index is >= the length of the list 
PASS document.createTouchList exists and correctly creates a TouchList from a single Touch 
PASS touch list is a TouchList object 
PASS TouchList.length attribute exists 
PASS TouchList.length attribute is of type number (unsigned long) 
PASS TouchList.item attribute exists 
PASS TouchList.item attribute is of type function 
PASS touch point is a Touch object 
PASS Touch.identifier attribute exists 
PASS Touch.identifier attribute is of type number (long) 
PASS Touch.target attribute exists 
PASS Touch.target attribute is of type Element 
PASS Touch.screenX attribute exists 
PASS Touch.screenX attribute is of type number (long) 
PASS Touch.screenY attribute exists 
PASS Touch.screenY attribute is of type number (long) 
PASS Touch.clientX attribute exists 
PASS Touch.clientX attribute is of type number (long) 
PASS Touch.clientY attribute exists 
PASS Touch.clientY attribute is of type number (long) 
PASS Touch.pageX attribute exists 
PASS Touch.pageX attribute is of type number (long) 
PASS Touch.pageY attribute exists 
PASS Touch.pageY attribute is of type number (long) 
PASS TouchList.item returns null if the index is >= the length of the list 
PASS document.createTouchList exists and correctly creates a TouchList from two Touch objects 
PASS touch list is a TouchList object 
PASS TouchList.length attribute exists 
PASS TouchList.length attribute is of type number (unsigned long) 
PASS TouchList.item attribute exists 
PASS TouchList.item attribute is of type function 
PASS touch point is a Touch object 
PASS Touch.identifier attribute exists 
PASS Touch.identifier attribute is of type number (long) 
PASS Touch.target attribute exists 
PASS Touch.target attribute is of type Element 
PASS Touch.screenX attribute exists 
PASS Touch.screenX attribute is of type number (long) 
PASS Touch.screenY attribute exists 
PASS Touch.screenY attribute is of type number (long) 
PASS Touch.clientX attribute exists 
PASS Touch.clientX attribute is of type number (long) 
PASS Touch.clientY attribute exists 
PASS Touch.clientY attribute is of type number (long) 
PASS Touch.pageX attribute exists 
PASS Touch.pageX attribute is of type number (long) 
PASS Touch.pageY attribute exists 
PASS Touch.pageY attribute is of type number (long) 
PASS touch point is a Touch object 
PASS Touch.identifier attribute exists 
PASS Touch.identifier attribute is of type number (long) 
PASS Touch.target attribute exists 
PASS Touch.target attribute is of type Element 
PASS Touch.screenX attribute exists 
PASS Touch.screenX attribute is of type number (long) 
PASS Touch.screenY attribute exists 
PASS Touch.screenY attribute is of type number (long) 
PASS Touch.clientX attribute exists 
PASS Touch.clientX attribute is of type number (long) 
PASS Touch.clientY attribute exists 
PASS Touch.clientY attribute is of type number (long) 
PASS Touch.pageX attribute exists 
PASS Touch.pageX attribute is of type number (long) 
PASS Touch.pageY attribute exists 
PASS Touch.pageY attribute is of type number (long) 
PASS TouchList.item returns null if the index is >= the length of the list 
